Las gráficas ya no se rompen
